Move connection related code to a separate module
