+++ src/lib.rs
@@ -31,6 +31,9 @@ extern crate hamcrest2;
 #[macro_use]
 extern crate maplit;
 
+use quick_error::quick_error;
+
+mod connect;
 mod listener;
 mod message;
 mod peer;
@@ -39,13 +42,13 @@ mod priv_prelude;
 #[macro_use]
 mod utils;
 
+pub use crate::connect::{connect_first_ok, ConnectError, Connection, ConnectionError};
 pub use crate::listener::ConnectionListener;
 pub use crate::message::Message;
-pub use crate::peer::{connect_first_ok, ConnectError, Connection, ConnectionError, PeerInfo};
+pub use crate::peer::PeerInfo;
 pub use crate::peer_discovery::{discover_peers, shout_for_peers, DiscoveryError, DiscoveryServer};
 
 use maidsafe_utilities::serialisation::SerialisationError;
-use quick_error::quick_error;
 use std::io;
 
 // TODO(povilas): use failure crate
